Decode updated for branch with links Execute updated with a lot of cases Register getLong() fixed ControlUnit setFlag() written, but not used.

+++ src/simulation/registers/Register.java
@@ -326,7 +326,7 @@ public class Register {
             return null;
         }
 
-        long word = ByteBuffer.wrap(Arrays.copyOfRange(bytes, startByte, startByte + Long.BYTES)).getInt();
+        long word = ByteBuffer.wrap(Arrays.copyOfRange(bytes, startByte, startByte + Long.BYTES)).getLong();
         return word;
     }
